+++ scripts/container-build.sh
@@ -220,6 +220,12 @@ MULTILIB_FLAGS=""
 BINUTILS_PATCH=""
 GDB_PATCH=""
 
+WITH_GDB_PY="y"
+WITH_GDB_PY3=""
+USE_PLATFORM_PYTHON=""
+USE_PLATFORM_PYTHON3=""
+PYTHON3_VERSION=""
+
 # -----------------------------------------------------------------------------
 
 # Redefine to "y" to create the LTO plugin links.
@@ -246,7 +252,146 @@ README_OUT_FILE_NAME="README-${RELEASE_VERSION}.md"
 
 # In reverse chronological order.
 # Keep them in sync with https://github.com/sifive/freedom-tools/releases.
-if [[ "${RELEASE_VERSION}" =~ 8\.2\.0-3\.* ]]
+if [[ "${RELEASE_VERSION}" =~ 8\.3\.0-* ]]
+then
+  # This is similar to SiFive 2019.08.0 release.
+  # https://github.com/sifive/freedom-tools/releases
+
+  # Binutils 2.32 with SiFive CLIC patches
+  # https://github.com/sifive/riscv-binutils-gdb/tree/03d23d58701bdd425c613b0be2d458bcde46912b
+
+  # GCC 8.3.0 with SiFive CLIC patches
+  # https://github.com/sifive/riscv-gcc/tree/e195042babe2dc30d9fabb88d336d8c8679b3702
+
+  # Newlib 3.1.0 from SiFive branch
+  # https://github.com/sifive/riscv-newlib/tree/0d24a86822a5ee73d6a6aa69e2a0118aa1e35204
+
+  # GDB 8.3 from FSF gdb-8.3-release branch
+  # riscv-gdb @ 9b40759 (11 May 2019)
+  # https://sourceware.org/git/?p=binutils-gdb.git
+  # git://sourceware.org/git/binutils-gdb.git
+
+
+  # ---------------------------------------------------------------------------
+
+  # Inspired from SiFive
+  # MULTILIBS_GEN :=            rv32e-ilp32e--c rv32em-ilp32e--c rv32eac-ilp32e-- rv32emac-ilp32e-- rv32i-ilp32--c rv32im-ilp32--c rv32imf-ilp32f--c rv32iac-ilp32-- rv32imac-ilp32-- rv32imafc-ilp32f-rv32imafdc- rv32imafdc-ilp32d-- rv64i-lp64--c rv64im-lp64--c rv64imf-lp64f--c rv64iac-lp64-- rv64imac-lp64-- rv64imafc-lp64f-rv64imafdc- rv64imafdc-lp64d--
+
+  # Minimal list, for tests only. Pass it via the environment.
+  # GCC_MULTILIB=${GCC_MULTILIB:-"rv32imac-ilp32-- rv64imac-lp64--"}
+
+  # New extended list, based on SiFive list.
+  # Added: rv32imaf-ilp32f--
+  GCC_MULTILIB=${GCC_MULTILIB:-"\
+  rv32e-ilp32e--c \
+	rv32ea-ilp32e--m \
+	rv32em-ilp32e--c \
+	rv32eac-ilp32e-- \
+	rv32emac-ilp32e-- \
+	rv32i-ilp32--c \
+	rv32ia-ilp32--m \
+	rv32im-ilp32--c \
+	rv32if-ilp32f-rv32ifd-c \
+	rv32iaf-ilp32f-rv32imaf,rv32iafc-d \
+	rv32imf-ilp32f-rv32imfd-c \
+  rv32imaf-ilp32f-- \
+	rv32iac-ilp32-- \
+	rv32imac-ilp32-- \
+	rv32imafc-ilp32f-rv32imafdc- \
+	rv32ifd-ilp32d--c \
+	rv32imfd-ilp32d--c \
+	rv32iafd-ilp32d-rv32imafd,rv32iafdc- \
+	rv32imafdc-ilp32d-- \
+	rv64i-lp64--c \
+	rv64ia-lp64--m \
+	rv64im-lp64--c \
+	rv64if-lp64f-rv64ifd-c \
+	rv64iaf-lp64f-rv64iafc-d \
+	rv64imf-lp64f-rv64imfd-c \
+	rv64imaf-lp64f-- \
+	rv64iac-lp64-- \
+	rv64imac-lp64-- \
+	rv64imafc-lp64f-rv64imafdc- \
+	rv64ifd-lp64d--m,c \
+	rv64iafd-lp64d-rv64imafd,rv64iafdc- \
+	rv64imafdc-lp64d-- \
+"}
+
+  GCC_MULTILIB_FILE=${GCC_MULTILIB_FILE:-"t-elf-multilib"}
+
+  # ---------------------------------------------------------------------------
+
+  BINUTILS_VERSION="2.32"
+  # From gcc/BASE_VER
+  GCC_VERSION="8.3.0"
+  # From newlib/configure, VERSION=
+  NEWLIB_VERSION="3.1.0"
+  # From gdb/VERSION.in
+  GDB_VERSION="8.3"
+
+  # ---------------------------------------------------------------------------
+
+  if [ "${USE_GITS}" != "y" ]
+  then
+
+    # Be sure there is no `v`, it is added in the URL.
+    GH_RELEASE="8.3.0-1.1"
+    BINUTILS_GH_RELEASE=${BINUTILS_GH_RELEASE:-"${GH_RELEASE}"}
+    GCC_GH_RELEASE=${GCC_GH_RELEASE:-"${GH_RELEASE}"}
+    NEWLIB_GH_RELEASE=${NEWLIB_GH_RELEASE:-"${GH_RELEASE}"}
+    # Same, with a `-gdb` suffix added.
+    GDB_GH_RELEASE=${GDB_GH_RELEASE:-"${GH_RELEASE}-gdb"}
+
+  else
+
+    BINUTILS_GIT_BRANCH=${BINUTILS_GIT_BRANCH:-"sifive-binutils-2.32-xpack"}
+    # 16 April 2019
+    BINUTILS_GIT_COMMIT=${BINUTILS_GIT_COMMIT:-"03d23d58701bdd425c613b0be2d458bcde46912b"}
+
+    GCC_GIT_BRANCH=${GCC_GIT_BRANCH:-"sifive-gcc-8.3.0-xpack"}
+    GCC_GIT_COMMIT=${GCC_GIT_COMMIT:-"e195042babe2dc30d9fabb88d336d8c8679b3702"}
+
+    NEWLIB_GIT_BRANCH=${NEWLIB_GIT_BRANCH:-"sifive-master-xpack"}
+    NEWLIB_GIT_COMMIT=${NEWLIB_GIT_COMMIT:-"0d24a86822a5ee73d6a6aa69e2a0118aa1e35204"}
+
+    GDB_SRC_FOLDER_NAME=${GDB_SRC_FOLDER_NAME:-"binutils-gdb.git"}
+    GDB_GIT_BRANCH=${GDB_GIT_BRANCH:-"sifive-gdb-8.3-xpack"}
+    GDB_GIT_COMMIT=${GDB_GIT_COMMIT:-"9b8cecd18313807ac0cc4d2b1871603279b94244"}
+
+  fi
+  
+  # ---------------------------------------------------------------------------
+
+  ZLIB_VERSION="1.2.8"
+  GMP_VERSION="6.1.2"
+  MPFR_VERSION="3.1.6"
+  MPC_VERSION="1.0.3"
+  ISL_VERSION="0.18"
+  LIBELF_VERSION="0.8.13"
+  EXPAT_VERSION="2.2.5"
+  LIBICONV_VERSION="1.15"
+  XZ_VERSION="5.2.3"
+
+  PYTHON_WIN_VERSION="2.7.13"
+
+  # GDB 8.3 with Python3 not yet functional on Windows.
+  # GDB does not know the Python3 API when compiled with mingw.
+  if [ "${TARGET_PLATFORM}" != "win32" ]
+  then
+    # checking for python3.7... no
+    # configure: error: no usable python found at /Users/ilg/opt/xbb/bin/python3
+
+    : # WITH_GDB_PY3="y" 
+    : # PYTHON3_VERSION="3.7.2"
+  fi
+
+  USE_PLATFORM_PYTHON="y"
+
+  BINUTILS_PATCH="binutils-gdb-${BINUTILS_VERSION}.patch"
+  GDB_PATCH="binutils-gdb-${BINUTILS_VERSION}.patch"
+
+  # ---------------------------------------------------------------------------
+elif [[ "${RELEASE_VERSION}" =~ 8\.2\.0-3\.* ]]
 then
   # This is similar to SiFive 2019.05.0 release.
   # https://github.com/sifive/freedom-tools/releases
@@ -455,7 +600,15 @@ echo
 if [ "${TARGET_PLATFORM}" == "win32" ]
 then
   # The Windows GDB needs some headers from the Python distribution.
-  download_python_win
+  if [ "${WITH_GDB_PY}" == "y" ]
+  then
+    download_python_win
+  fi
+
+  if [ "${WITH_GDB_PY3}" == "y" ]
+  then
+    download_python3_win
+  fi
 fi
 
 # -----------------------------------------------------------------------------
@@ -516,9 +669,15 @@ fi
 # Task [III-6] /$HOST_NATIVE/gdb/
 # Task [IV-4] /$HOST_MINGW/gdb/
 do_gdb ""
-do_gdb "-py"
-# Python3 support not yet functional.
-# do_gdb "-py3"
+if [ "${WITH_GDB_PY}" == "y" ]
+then
+  do_gdb "-py"
+fi
+
+if [ "${WITH_GDB_PY3}" == "y" ]
+then
+  do_gdb "-py3"
+fi
 
 # Task [III-7] /$HOST_NATIVE/build-manual
 # Nope, the build process is different.
@@ -578,7 +737,15 @@ run_gdb
 
 if [  "${TARGET_PLATFORM}" != "win32" ]
 then
-  run_gdb "-py"
+  if [ "${WITH_GDB_PY}" == "y" ]
+  then
+    run_gdb "-py"
+  fi
+
+  if [ "${WITH_GDB_PY3}" == "y" ]
+  then
+    run_gdb "-py3"
+  fi
 fi
 
 # -----------------------------------------------------------------------------